Bug fix and update for v1.1.2
1) Corrected error when calculating percentiles

2) Fixed bug where the EM colors were not correct when plotting in
MATLAB 2014b and higher

3) Fixed a bug where specimen names beginning with a number were not
properly displayed above the data plot or when the figure was exported.

4) Adjusted lower search bound of SGG q parameter from -1 to -0.25.
This ensures better fitting.

5) Fixed occasional bug when getting parametric EM initial guess. This
was related to a lack of stationary points in some grain size spectra.

6) Moved the projection to simplex validity check into the optimized
function. This can slow down the speed of paramedic fitting, but
ensures better fits.

7) Added function for force closing all AnalySize windows (used for
development and debugging).